requirements.txt
python-telegram-bot
requests
xmltodict
beautifulsoup4
Django